refactor ScrollbarBridge to use MacScrollbarHelper
