+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/ScrollbarBridge.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.takeOrElse
 import androidx.compose.ui.unit.dp
 import com.intellij.ui.mac.foundation.Foundation
+import org.jetbrains.jewel.bridge.MacScrollbarHelper
 import org.jetbrains.jewel.bridge.retrieveColorOrUnspecified
 import org.jetbrains.jewel.ui.component.styling.ScrollbarColors
 import org.jetbrains.jewel.ui.component.styling.ScrollbarMetrics
@@ -26,7 +27,7 @@ internal fun readScrollbarStyle(isDark: Boolean): ScrollbarStyle =
 
 private fun readScrollbarVisibility() =
     if (hostOs.isMacOS) {
-        readMacScrollbarStyle()
+        MacScrollbarHelper.scrollbarVisibility
     } else {
         ScrollbarVisibility.AlwaysVisible
     }
@@ -40,7 +41,7 @@ private fun readScrollbarColors(isDark: Boolean) =
 
 private fun readTrackClickBehavior() =
     if (hostOs.isMacOS) {
-        readMacScrollbarBehavior()
+        MacScrollbarHelper.trackClickBehavior
     } else {
         TrackClickBehavior.JumpToSpot
     }
@@ -193,28 +194,6 @@ private fun readScrollbarMetrics(): ScrollbarMetrics =
             trackPaddingExpanded = PaddingValues(),
         )
     }
-
-private fun readMacScrollbarStyle(): ScrollbarVisibility {
-    val nsScroller =
-        Foundation
-            .invoke(Foundation.getObjcClass("NSScroller"), "preferredScrollerStyle")
-
-    val visibility: ScrollbarVisibility =
-        if (1 == nsScroller.toInt()) {
-            ScrollbarVisibility.WhenScrolling.Companion.defaults()
-        } else {
-            ScrollbarVisibility.AlwaysVisible
-        }
-    return visibility
-}
-
-private fun readMacScrollbarBehavior(): TrackClickBehavior {
-    val defaults = Foundation.invoke("NSUserDefaults", "standardUserDefaults")
-    Foundation.invoke(defaults, "synchronize")
-    return Foundation
-        .invoke(defaults, "boolForKey:", Foundation.nsString("AppleScrollerPagingBehavior"))
-        .run { if (toInt() == 1) TrackClickBehavior.JumpToSpot else TrackClickBehavior.NextPage }
-}
 
 public fun ScrollbarVisibility.WhenScrolling.Companion.defaults(
     appearAnimationDuration: Duration = 125.milliseconds,
